pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brev-launcher"
version = "0.1.0"
description = "Generate Brev Launchable config files from your project"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Brev Community" }
]
keywords = ["brev", "launchable", "gpu", "jupyter", "deployment"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "ruamel.yaml>=0.18.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
brev-launcher = "brev_launcher.cli:app"

[project.urls]
Homepage = "https://github.com/brevdev/brev-launcher"
Documentation = "https://github.com/brevdev/brev-launcher#readme"
Repository = "https://github.com/brevdev/brev-launcher"

[tool.hatch.build.targets.wheel]
packages = ["src/brev_launcher"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

